requirements.txt
discord.py==1.6.0
PyNaCl==1.4.0
